Bugfix/tests

+++ parser/xml/discriminator/_fulltext.py
@@ -1,4 +1,5 @@
 import array
+import math
 from lxml import etree
 
 
@@ -40,9 +41,9 @@ class _Fulltext(object):
         # calculating median
         if counter != 0:
             if counter % 2 == 0:
-                a = font_sizes[counter / 2]
-                b = font_sizes[(counter / 2) - 1]
+                a = font_sizes[int(counter / 2)]
+                b = font_sizes[int(counter / 2) - 1]
                 median = (a + b) / 2
             else:
-                median = font_sizes[counter / 2]
+                median = font_sizes[int(counter / 2)]
         return median
